processor/requirements.txt
psycopg2==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.0
geopandas==0.14.0
Shapely==2.0.1
numpy==1.24.3
python-dotenv==1.0.0
geopandas==0.14.0
pyproj==3.6.0
fiona==1.9.4
shapely==2.0.1
python-dotenv==1.0.0
ratelimit==2.2.1
sodapy==2.2.0  # For the Socrata Open Data API
beautifulsoup4==4.12.2
requests==2.31.0